+++ django_mysql/utils.py
@@ -9,7 +9,9 @@ from collections import defaultdict
 from contextlib import contextmanager
 from subprocess import PIPE, Popen, call
 from threading import Lock, Thread
+from weakref import WeakKeyDictionary
 
+from django.db import connection as default_connection
 from django.db import DEFAULT_DB_ALIAS, connections
 from django.utils import six
 from django.utils.six.moves.queue import Empty, Queue
@@ -94,10 +96,24 @@ def format_duration(total_seconds):
     return ''.join(out)
 
 
+_is_mariadb_cache = WeakKeyDictionary()
+
+
 def connection_is_mariadb(connection):
-    with connection.temporary_connection():
-        server_info = connection.connection.get_server_info()
-        return 'MariaDB' in server_info
+    if connection.vendor != 'mysql':
+        return False
+
+    if connection is default_connection:
+        connection = connections[DEFAULT_DB_ALIAS]
+
+    try:
+        return _is_mariadb_cache[connection]
+    except KeyError:
+        with connection.temporary_connection():
+            server_info = connection.connection.get_server_info()
+        is_mariadb = 'MariaDB' in server_info
+        _is_mariadb_cache[connection] = is_mariadb
+        return is_mariadb
 
 
 def settings_to_cmd_args(settings_dict):